updated ReceptHandler.java

+++ src/main/java/Handlers/ReceptHandler.java
@@ -189,7 +189,7 @@ public class ReceptHandler implements IReceptHandler {
             while (resultSet.next()){
                 receptKomponentList.add(new ReceptKomponentDTO(
                         resultSet.getInt("fk_RaavareID"),
-                        resultSet.getInt("nonNetto"),
+                        resultSet.getDouble("nonNetto"),
                         resultSet.getDouble("tolerance"),
                         resultSet.getInt("fk_ReceptID")
                     )
